refactor(editor): Try to reduce down bundle size of Serlo frontend by moving folders
