+++ packages/editor/src/core/index.ts
@@ -1,7 +1,6 @@
 export { SubDocument } from './sub-document'
 export type { SubDocumentProps } from './sub-document'
 export { Editor } from './editor'
-export { AllPlugins } from './all-plugins'
 export type { EditorProps, BaseEditor } from './types'
 export { PreferenceContext } from './contexts'
 export type { Preference } from './contexts'
