+++ packages/editor/src/package/index.ts
@@ -2,7 +2,7 @@ export { SerloEditor, type SerloEditorProps } from './editor'
 export { SerloRenderer, type SerloRendererProps } from './serlo-renderer'
 
 export type { BaseEditor } from '@editor/core'
-export { AllPlugins } from '@editor/core'
+export { AllPlugins } from './all-plugins'
 
 export { EditorPluginType } from '@editor/types/editor-plugin-type'
